MELQ-150: Move PostEdit to the separate directory
